Make settings overrides more sane.
This simplifies the way settings are stored and retrieved.

The config `options` function now handles ftype specifics.

When a profile is applied, it is now merged into the sessionOverrides table, intead of being applied on top of it.

That means it is now possible to override e.g. `Command` after a profile has been applied.

If a profile is applied on top of it again, the overriden `Command` is now wiped.